+++ racksdb/generic/dumpers/json.py
@@ -8,15 +8,16 @@ from typing import Any
 import json
 import logging
 
-from ...generic.db import DBObject, DBObjectRange, DBObjectRangeId, DBDict
+from ...generic.db import DBObject, DBObjectRange, DBObjectRangeId, DBDict, DBList
 
 logger = logging.getLogger(__name__)
 
 
 class GenericJSONEncoder(json.JSONEncoder):
-    def __init__(self, objects_map={}, **kwargs):
+    def __init__(self, objects_map={}, fold=True, **kwargs):
         super().__init__(**kwargs)
         self.objects_map = objects_map
+        self.fold = fold
 
     def default(self, obj: Any) -> Any:
         if isinstance(obj, DBObjectRange):
@@ -24,7 +25,22 @@ class GenericJSONEncoder(json.JSONEncoder):
         elif isinstance(obj, DBObjectRangeId):
             return obj.start
         elif isinstance(obj, DBDict):
-            return [self.default(item) for item in obj.values()]
+            if self.fold:
+                # Force iteration over the values of the dictionnary to avoid automatic
+                # expansion performed by DBDict iterator.
+                return [item for item in obj.values()]
+            else:
+                # Use DBDict iterator to expand potential DBExpandableObject.
+                return [item for item in obj]
+        elif isinstance(obj, DBList):
+            if self.fold:
+                return [item for item in obj.itervalues()]
+            else:
+                # As a DBList is also a standard iterable Python list, json.JSONEncoder
+                # will iterate itself over the DBList, thus expanding potential
+                # expandable objects automatically. There is nothing more to do in this
+                # case.
+                return obj
         elif isinstance(obj, DBObject):
             result = {}
             for attribute, value in vars(obj).items():
@@ -50,7 +66,11 @@ class GenericJSONEncoder(json.JSONEncoder):
                         continue
                     # Else, map the object to one of its attribute.
                     value = getattr(value, self.objects_map[value.__class__.__name__])
-                if isinstance(value, DBObject) or isinstance(value, DBDict):
+                if (
+                    isinstance(value, DBObject)
+                    or isinstance(value, DBDict)
+                    or isinstance(value, DBList)
+                ):
                     result[attribute] = self.default(value)
                 else:
                     result[attribute] = value
@@ -60,8 +80,9 @@ class GenericJSONEncoder(json.JSONEncoder):
 
 
 class DBDumperJSON:
-    def __init__(self, show_types=False, objects_map={}):
+    def __init__(self, show_types=False, objects_map={}, fold=True):
         self.objects_map = objects_map
+        self.fold = fold
 
     def dump(self, obj: Any) -> str:
         # DBDict are also standard python dictionnaries, then json.JSONEncoder thinks it
@@ -76,5 +97,10 @@ class DBDumperJSON:
         #
         # For these reasons, DBDict is converted here as a standard list of values.
         if isinstance(obj, DBDict):
-            obj = [item for item in obj.values()]
-        return GenericJSONEncoder(objects_map=self.objects_map).encode(obj)
+            if self.fold:
+                obj = [item for item in obj.values()]
+            else:
+                obj = [item for item in obj]
+        return GenericJSONEncoder(objects_map=self.objects_map, fold=self.fold).encode(
+            obj
+        )